fix: two real defects, both found by READING — neither was tool-flagged
A broad clang-tidy + splint sweep over the hand-authored C and Clojure, mined
by six agents. 1950 clang-tidy findings in our own C, of which 1672 (86%) were
five style-policy checks; 55 splint findings. Of the ~204 plausible-signal
findings actually adjudicated, the tools' real-defect yield was ZERO — every
one was correctly dismissed, several with better reasoning than the check
deserved (the narrowing-conversion hits in chart_fade_draw_cb turn out to be a
byte-identical port of the vendored LVGL demo's own cast, and the
math-missing-parentheses hits mirror LVGL's update_knob_pos verbatim, which
the function comment says is deliberate).

The two defects below came from agents READING the code around the noise. Both
verified by hand here against the source before being touched.

1. gesture.c — SIGNED OVERFLOW in the double-tap window (undefined behaviour).
   PointerEvent.event_time is `uint64` carrying the W3C event.timeStamp: ms
   since navigation start, monotonically increasing for the life of the host
   session. main.c narrows it to int32 (`.t_ms = (int32_t)t_ms`, and again in
   pointer_release_cancel), so it wraps after ~2^31 ms — about 24.8 days of
   uptime, which an always-on ARM kiosk reaches.

   classify_release then computed `s->t_ms - g->last_tap_t` on two int32
   values. For a tap pair straddling that wrap the subtraction overflows, and
   signed overflow is UB the -O2 build is entitled to assume away. Now computed
   in uint32, where wraparound is DEFINED and yields the correct elapsed ms for
   any gap under 2^31 — i.e. every gap a double-tap window can care about.

   Not caught by bugprone-narrowing-conversions, because the narrowing is an
   EXPLICIT cast and the check only flags implicit ones. A cast is how you tell
   the compiler you meant it; it is not evidence that you did.

2. binary_dedup.clj — a fail-closed gate that failed OPEN at the top level.
   `has-map-field?` throws :unresolved-type-reference when a NESTED field's
   type is missing from the descriptor. `validate-determinism!` used `when-let`
   for the TOP-LEVEL lookup, so an unresolvable subsystem type silently skipped
   the field and the doseq moved on as if it had passed.

   The docstring promised it "Throws ... if any of the fail-closed cases in
   has-map-field? trigger" — a promise it could not keep, because when the
   top-level type fails to resolve has-map-field? is never called at all. The
   namespace docstring calls this a fail-closed whitelist; it now is one. No
   test covered the path: the "every fail-closed rejection path" suite feeds
   unresolved refs into has-map-field? directly, never through this branch.

Supporting change — renderer/wasm.mk emits compile_commands.json (`compile-db`)
from the build's OWN variables. This exists because a hand-assembled flag list
disagrees with the build silently, and clang-tidy then reports diagnostics the
compiler never sees. The generator promptly proved its own point: the first
version omitted -std=c23, which the app rule adds, producing three phantom
parse errors on a `static_assert` that compiles cleanly, and inflating
bugprone-macro-parentheses from 6 to 18. `APP_STD` is now one variable used by
both the compile rule and the db target, so the two cannot drift again.

PROOF: protodoc 226 tests / 21561 assertions / 0 failures / 0 errors; renderer
battery GREEN with demo-parity 3 of 3 bit-equal. All four gates green.

CONSEQUENCES — consumers: REGENERATE. controls.wasm changed (gesture.c is
compiled into it). The double-tap window now behaves correctly across the
~24.8-day event_time wrap instead of relying on UB; no API, wire, or pixel
change — goldens and gallery are byte-unchanged.

+++ renderer/src/gesture.c
@@ -146,8 +146,16 @@ int32_t gesture_on_move(gesture_recognizer_t *g, const gesture_sample_t *s,
 static int32_t classify_release(const gesture_recognizer_t *g,
                                 const gesture_sample_t *s,
                                 gesture_decision_t *out) {
+  /* UNSIGNED delta on purpose. t_ms carries the host's W3C event.timeStamp
+   * (ms since navigation start) narrowed to int32 by the shell, so it wraps
+   * after ~24.8 days of session uptime. Subtracting two int32 values that
+   * straddle that wrap is SIGNED OVERFLOW — undefined behaviour, which -O2 is
+   * entitled to assume cannot happen. Computed in uint32 the wraparound is
+   * well-defined and yields the correct elapsed ms for any gap under 2^31,
+   * which is every gap a double-tap window could care about. */
   if (g->has_last_tap &&
-      (double)(s->t_ms - g->last_tap_t) <= (double)GESTURE_DOUBLE_TAP_MS &&
+      (double)((uint32_t)s->t_ms - (uint32_t)g->last_tap_t) <=
+          (double)GESTURE_DOUBLE_TAP_MS &&
       gesture_ndc_dist(g->last_tap_x, g->last_tap_y, s->x, s->y) <=
           px_to_ndc((double)GESTURE_DOUBLE_TAP_PX)) {
     out[0].kind = GESTURE_KIND_TRACK;
